third_party/blink/web_tests/external/wpt/css/css-text-decor/parsing/text-emphasis-position-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-emphasis-position'] = "auto" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-emphasis-position'] = "right under" should set the property value
  assert_equals: serialization should be canonical expected "under" but got "under right"
Harness: the test ran to completion.
